Moved ExchnageBuffer stuff into one file to comply with convention
